separated findNoteIndexAndNullifier
